moved test data downloading to tests.utils.py and started test for fit_regressors
