+++ tedana/tests/test_integration.py
@@ -1,24 +1,19 @@
 """Integration tests for "real" data."""
 
 import glob
-import json
 import logging
 import os
 import os.path as op
 import re
 import shutil
 import subprocess
-import tarfile
-from datetime import datetime
-from gzip import GzipFile
-from io import BytesIO
 
 import pandas as pd
 import pytest
-import requests
 from pkg_resources import resource_filename
 
 from tedana.io import InputHarvester
+from tedana.tests.utils import data_for_testing_info, download_test_data
 from tedana.workflows import t2smap as t2smap_cli
 from tedana.workflows import tedana as tedana_cli
 from tedana.workflows.ica_reclassify import ica_reclassify_workflow
@@ -74,129 +69,6 @@ def check_integration_outputs(fname, outpath, n_logs=1):
             msg += "\nFound but not expected:\n\t"
             msg += "\n\t".join(found_not_expected)
         raise ValueError(msg)
-
-
-def data_for_testing_info(test_dataset=str):
-    """
-    Get the path and download link for each dataset used for testing.
-
-    Also creates the base directories into which the data and output
-    directories are written
-
-    Parameters
-    ----------
-    test_dataset : str
-       References one of the datasets to download. It can be:
-        three-echo
-        three-echo-reclassify
-        four-echo
-        five-echo
-
-    Returns
-    -------
-    test_data_path : str
-       The path to the local directory where the data will be downloaded
-    osf_id : str
-       The ID for the OSF file.
-       Data download link would be https://osf.io/osf_id/download
-       Metadata download link would be https://osf.io/osf_id/metadata/?format=datacite-json
-    """
-
-    tedana_path = os.path.dirname(tedana_cli.__file__)
-    base_data_path = os.path.abspath(os.path.join(tedana_path, "../../.testing_data_cache"))
-    os.makedirs(base_data_path, exist_ok=True)
-    os.makedirs(os.path.join(base_data_path, "outputs"), exist_ok=True)
-    if test_dataset == "three-echo":
-        test_data_path = os.path.join(base_data_path, "three-echo/TED.three-echo")
-        osf_id = "rqhfc"
-        os.makedirs(os.path.join(base_data_path, "three-echo"), exist_ok=True)
-        os.makedirs(os.path.join(base_data_path, "outputs/three-echo"), exist_ok=True)
-    elif test_dataset == "three-echo-reclassify":
-        test_data_path = os.path.join(base_data_path, "reclassify")
-        osf_id = "f6g45"
-        os.makedirs(os.path.join(base_data_path, "outputs/reclassify"), exist_ok=True)
-    elif test_dataset == "four-echo":
-        test_data_path = os.path.join(base_data_path, "four-echo/TED.four-echo")
-        osf_id = "gnj73"
-        os.makedirs(os.path.join(base_data_path, "four-echo"), exist_ok=True)
-        os.makedirs(os.path.join(base_data_path, "outputs/four-echo"), exist_ok=True)
-    elif test_dataset == "five-echo":
-        test_data_path = os.path.join(base_data_path, "five-echo/TED.five-echo")
-        osf_id = "9c42e"
-        os.makedirs(os.path.join(base_data_path, "five-echo"), exist_ok=True)
-        os.makedirs(os.path.join(base_data_path, "outputs/five-echo"), exist_ok=True)
-    else:
-        raise ValueError(f"{test_dataset} is not a valid dataset string for data_for_testing_info")
-
-    return test_data_path, osf_id
-
-
-def download_test_data(osf_id, test_data_path):
-    """If current data is not already available, downloads tar.gz data.
-
-    Data are stored at `https://osf.io/osf_id/download`.
-    It unpacks into `out_path`.
-
-    Parameters
-    ----------
-    osf_id : str
-       The ID for the OSF file.
-    out_path : str
-        Path to directory where OSF data should be extracted
-    """
-
-    try:
-        datainfo = requests.get(f"https://osf.io/{osf_id}/metadata/?format=datacite-json")
-    except Exception:
-        if len(os.listdir(test_data_path)) == 0:
-            raise ConnectionError(
-                f"Cannot access https://osf.io/{osf_id} and testing data " "are not yet downloaded"
-            )
-        else:
-            TestLGR.warning(
-                f"Cannot access https://osf.io/{osf_id}. "
-                f"Using local copy of testing data in {test_data_path} "
-                "but cannot validate that local copy is up-to-date"
-            )
-            return
-    datainfo.raise_for_status()
-    metadata = json.loads(datainfo.content)
-    # 'dates' is a list with all udpates to the file, the last item in the list
-    # is the most recent and the 'date' field in the list is the date of the last
-    # update.
-    osf_filedate = metadata["dates"][-1]["date"]
-
-    # File the file with the most recent date for comparision with
-    # the lsst updated date for the osf file
-    if os.path.exists(test_data_path):
-        filelist = glob.glob(f"{test_data_path}/*")
-        most_recent_file = max(filelist, key=os.path.getctime)
-        if os.path.exists(most_recent_file):
-            local_filedate = os.path.getmtime(most_recent_file)
-            local_filedate_str = str(datetime.fromtimestamp(local_filedate).date())
-            local_data_exists = True
-        else:
-            local_data_exists = False
-    else:
-        local_data_exists = False
-    if local_data_exists:
-        if local_filedate_str == osf_filedate:
-            TestLGR.info(
-                f"Downloaded and up-to-date data already in {test_data_path}. Not redownloading"
-            )
-            return
-        else:
-            TestLGR.info(
-                f"Downloaded data in {test_data_path} was last modified on "
-                f"{local_filedate_str}. Data on https://osf.io/{osf_id} "
-                f" was last updated on {osf_filedate}. Deleting and redownloading"
-            )
-            shutil.rmtree(test_data_path)
-    req = requests.get(f"https://osf.io/{osf_id}/download")
-    req.raise_for_status()
-    t = tarfile.open(fileobj=GzipFile(fileobj=BytesIO(req.content)))
-    os.makedirs(test_data_path, exist_ok=True)
-    t.extractall(test_data_path)
 
 
 def reclassify_raw() -> str:
@@ -393,13 +265,12 @@ def test_integration_three_echo_external_regressors_single_model(skip_integratio
     #  Col 1 (trans_x_correlation) is the TS for ICA comp 59 + similar stdev Gaussian Noise
     #  Col 2 (trans_y_correlation) is 0.4*comp29+0.5+comp20+Gaussian Noise
     #  Col 3 (trans_z_correlation) is comp20+Gaussian Noise
-    # The above are the same as the test that uses corr for external regressors
     #  Col 4-6 are Gaussian noise representing pitch/roll/yaw
     #  Col 7-12 are the first derivative of col 1-6
-    # With the currently set up decision tree minimal_exteral 2, one component (my 59)
-    # should be rejected, but 20 and 29 aren't rejected because neither crosses the
-    # r>0.8 threshold. If trans_y and trans_Z were included in a single model then
-    # component 20 would have been rejected
+    # With the currently set up decision tree,
+    # Component 59 should be rejected because it is correlated to trans_x and,
+    # comp 20 should be rejected because of a signif fit to a combination of trans_y and trans_z.
+    # Component 29 is not rejected because the fit does not cross a r>0.8 threshold
     # Note that the above is in comparision to the minimal decision tree
     # but the integration test for 3 echoes uses the kundu tree
     download_test_data(osf_id, test_data_path)
